resources/populator/requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
happybase==1.2.0
idna==2.8
ply==3.11
pycparser==2.19
pycryptodome==3.8.2
regex==2019.6.8
requests==2.22.0
six==1.12.0
thriftpy2==0.4.4
urllib3==1.25.3
